Fix for OUJS
OUJS has made a change recently to require SPDX codes for OSI approved licensing.

In order to improve the appearance of your script homepages it would be appreciated if you could modify all of your affected scripts. Think this is all of them and hopefully in the right places.


Until this change is made you will be unable to update those affected scripts.

Thanks,
OUJS Staff